move custom exceptions into exceptions.py module
